firmware: dmi: handle missing DMI data gracefully
Currently, when booting a kernel with DMI support on a platform that has
no DMI tables, the following output is emitted into the kernel log:

  [    0.128818] DMI not present or invalid.
  ...
  [    1.306659] dmi: Firmware registration failed.
  ...
  [    2.908681] dmi-sysfs: dmi entry is absent.

The first one is a pr_info(), but the subsequent ones are pr_err()s that
complain about a condition that is not really an error to begin with.

So let's clean this up, and give up silently if dma_available is not set.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Acked-by: Martin Hundebøll <mnhu@prevas.dk>
Signed-off-by: Jean Delvare <jdelvare@suse.de>
Change-Id: I99d59350b07210e4a7eb4fcf839574ce50cda08f
Git-commit: a81114d03e4a529c4b68293249f75438b3c1783f
Git-repo: https://android.googlesource.com/kernel/common/
Signed-off-by: Manohar Vavilapalli <mvavilap@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>